cleanup blocks
